Moving api logger to server/internal
