+++ pkg/server/internal/logger.go
@@ -1,10 +1,11 @@
-package logger
+package internal
 
 import (
 	"fmt"
 
 	"github.com/go-logr/logr"
 	"github.com/go-logr/zapr"
+	"github.com/weaveworks/weave-gitops/pkg/logger"
 	"go.uber.org/zap"
 )
 
@@ -12,7 +13,7 @@ type ApiLogger struct {
 	logger logr.Logger
 }
 
-func NewApiLogger(z *zap.Logger) Logger {
+func NewApiLogger(z *zap.Logger) logger.Logger {
 	logger := zapr.NewLogger(z)
 
 	return ApiLogger{
